Add explicit int cast to getAmountDue calculation for consistency
Agent-Logs-Url: https://github.com/Bottelet/DaybydayCRM/sessions/0f9f225c-1173-42cb-8b98-1f65c799b97e

+++ app/Services/Invoice/InvoiceCalculator.php
@@ -58,7 +58,7 @@ class InvoiceCalculator
 
     public function getAmountDue()
     {
-        return new Money($this->getTotalPrice()->getAmount() - $this->invoice->payments()->sum('amount'));
+        return new Money((int) ($this->getTotalPrice()->getAmount() - $this->invoice->payments()->sum('amount')));
     }
 
     public function getInvoice()
